fix(rollback): config cycle import
