+++ client/app/rollback.go
@@ -3,11 +3,12 @@ package app
 import (
 	cmtcmd "github.com/cometbft/cometbft/cmd/cometbft/commands"
 	cmtcfg "github.com/cometbft/cometbft/config"
-	"github.com/piplabs/story/client/cmd"
+
+	"github.com/piplabs/story/client/config"
 	"github.com/piplabs/story/lib/errors"
 )
 
-func RollbackCometAndAppState(a *App, cometCfg cmtcfg.Config, rollbackCfg cmd.RollbackConfig) (lastHeight int64, lastHash []byte, err error) {
+func RollbackCometAndAppState(a *App, cometCfg cmtcfg.Config, rollbackCfg config.RollbackConfig) (lastHeight int64, lastHash []byte, err error) {
 	for range rollbackCfg.RollbackHeights {
 		lastHeight, lastHash, err = cmtcmd.RollbackState(&cometCfg, rollbackCfg.RemoveBlock)
 		if err != nil {
